Rename PeepData to PeepAnimationData
