+++ src/openrct2/peep/PeepAnimationData.cpp
@@ -7,7 +7,7 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
-#include "PeepData.h"
+#include "PeepAnimationData.h"
 
 #include "PeepSpriteIds.h"
