+++ src/openrct2/peep/PeepAnimationData.h
